refactor: modularize Tauri app structure with commands and app modules
